+++ src/app/(root)/(routes)/items/[itemId]/components/trade-section/SuggestList.tsx
@@ -1,28 +1,16 @@
 import SuggestCard from '@/components/domain/card/suggest-card'
 import { Tabs, TabsTrigger, TabsList, TabsContent } from '@/components/ui/Tabs'
-import { getSuggestions } from '@/services/suggest/suggest'
 import { ItemSuggestion } from '@/types'
 
 type SuggestListProps = {
-  itemId: string
+  suggestionData: []
 }
 
-async function getSuggestionsValue(itemId: string) {
-  try {
-    const res = await getSuggestions(itemId)
-    const data = await res.json()
-
-    return data.data.cardList
-  } catch (e) {
-    console.log(e)
-  }
-}
 /**
  * TODO : 스크롤바 디자인 수정
  * TODO : 실제 API 연결(useMutation 사용해서)
  */
-const SuggestList = async ({ itemId }: SuggestListProps) => {
-  const suggestions = await getSuggestionsValue(itemId)
+const SuggestList = async ({ suggestionData }: SuggestListProps) => {
   return (
     <Tabs defaultValue="offer">
       <TabsList>
@@ -35,7 +23,7 @@ const SuggestList = async ({ itemId }: SuggestListProps) => {
           value={type}
           className=" data-[state=inactive]:hidden h-[402px] overflow-y-auto"
         >
-          {suggestions
+          {suggestionData
             .filter((v: ItemSuggestion) => v.suggestionType === type)
             .map((v: ItemSuggestion) => (
               <SuggestCard
